+++ vm/devices/storage/disklayer_sqlite/src/lib.rs
@@ -62,6 +62,7 @@
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]
 
+mod auto_cache;
 pub mod resolver;
 
 use anyhow::Context;
